requirements.txt
fastapi
jinja2
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
uvicorn
passlib
bcrypt==3.2.0
psycopg[binary]
cryptography
alembic
pandas
pymysql
pyodbc
oracledb
openai
mysql-connector-python

